댓글달기 후 댓글 리스트 다시 리로드 개발(진행중)

+++ src/main/java/com/jaeygun/board/board/service/ReplyService.java
@@ -21,5 +21,5 @@ public interface ReplyService {
      * @param boardUid
      * @return
      */
-    List<ReplyDTO> getRecentReplyList(int boardUid, Pageable pageable);
+    List<ReplyDTO> getRecentReplyList(long boardUid, Pageable pageable);
 }
